TestArchivalFromNonArchival is incompatible with `go test -race`
